pyproject.toml
[project]
name = "devops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.2",
]
